+++ .gitignore
@@ -445,5 +445,5 @@ tags
 # report sheets folder
 /report_sheets
 
-/sheets/data_transfer/data/
+/coding_sheets/data_transfer/data/
 # End of Customizations
