+++ coding_sheets/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class SheetsConfig(AppConfig):
-    name = "sheets"
+    name = "coding_sheets"
