+++ coding_sheets/extractor_script.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import json
 from django.conf import settings
-import sheets.coding_info as coding_info
+import coding_sheets.coding_info as coding_info
 
 pd.set_option("display.max_colwidth", 100)
 
@@ -9,7 +9,9 @@ pd.set_option("display.max_colwidth", 100)
 def read_coding_sheet(filename):
     """Read the workbook and extract data into different dataframes."""
     # excel file
-    file_name = settings.BASE_DIR + "/sheets/data_transfer/data/" + f"{filename}.xlsx"
+    file_name = (
+        settings.BASE_DIR + "/coding_sheets/data_transfer/data/" + f"{filename}.xlsx"
+    )
     xl_file = pd.ExcelFile(file_name)
 
     # list of coding names
